Create ANDEquation.py
The problem is as follows:

You are given a int[] A that contains exactly N+1 elements. Your task is to construct an AND-equation using each element of A exactly once. (That is, N of them will be on the left hand side of the AND-equation and the remaining one will be on the right hand side.) If this is possible, return the value of Y in this AND-equation. If no AND-equation can be constructed, return -1. (It can be shown that for each A there is at most one possible value Y, so the return value is always defined correctly.) 
  

Definition:
     
Class: ANDEquation 
Method: restoreY 
Parameters: int[] 
Returns: int 
Method signature: int restoreY(int[] A)